backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
plotly==5.16.0
openai==0.28.0
anthropic==0.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyMySQL==1.1.0
requests==2.31.0
scikit-learn==1.3.0
seaborn==0.12.2
openpyxl==3.1.2
xlrd==2.0.1